+++ app/models/waste_carriers_engine/edit_registration.rb
@@ -2,6 +2,7 @@
 
 module WasteCarriersEngine
   class EditRegistration < TransientRegistration
+    include CanCheckIfRegistrationTypeChanged
     include CanCopyDataFromRegistration
     include CanUseEditRegistrationWorkflow
     include CanUseLock
